Refactor mainmenu token
