+++ packages/mainmenu/src/index.ts
@@ -11,3 +11,4 @@ export * from './run';
 export * from './settings';
 export * from './view';
 export * from './tabs';
+export * from './tokens';
