+++ packages/statusbar-extension/src/index.ts
@@ -37,7 +37,7 @@ import {
 
 import { ISettingRegistry } from '@jupyterlab/coreutils';
 
-import { IMainMenu } from '@jupyterlab/mainmenu';
+import { IMainMenu } from '@jupyterlab/mainmenu/lib/tokens';
 
 import { Title, Widget } from '@phosphor/widgets';
